Implemented FeatureClassifierMixin for ProtoPNet (#45)
* Implemented FeatureClassifierMixin for ProtoPNet

common.models
- Added relu activations to mnist_encoder and mnist_decoder

cpvae.cpvae_test
- Checked the output directory exists before we try to delete it

prototype.model
- Implemented the FeatureClassifier Mixin in the most straightforward way

Added prototype.model_test

prototype.networks:
- Switched leaky_relu to relu because something is broken in vectorizing the gradient calculation

* Small changes to prototype model and test

+++ pyroclast/cpvae/cpvae_test.py
@@ -1,4 +1,7 @@
+import os.path as osp
+
 from absl.testing import parameterized
+
 from pyroclast.common.tf_util import setup_tfds
 from pyroclast.cpvae.cpvae import learn
 
@@ -32,4 +35,5 @@ class CpVAETest(parameterized.TestCase):
 
     def tearDown(self):
         import shutil
-        shutil.rmtree(self.args['output_dir'])
+        if osp.exists(self.args['output_dir']):
+            shutil.rmtree(self.args['output_dir'])
